refactor: made the lineshape code more performant
Based on tests, the way that the lineshape simulation is performed for
Gaussians is as good as it gets without doing approximations: we use a
struct of arrays for summing multiple Gaussians, and uses `mapreduce` to
minimize allocations needed.

There are probably speedups to be had more generally by using Fast Gauss
transform algorithms on a whole, but not worth it at the current time.