feat: performant Gaussian sum
Realized that the previous implementation did not produce correct
results, which was also why it was so fast.

In this current implementation, we use einsum notation to perform the
reduction performantly. As such, there's a specialized version of the
functor for `MultiGaussians` that operates on vector input. The idea is
to use multiple dispatch to determine the optimal function to use.