feat: added PGopher linelist parser
Implemented a CSV reader for parsing PGopher linelists: point to a CSV
file, and it will load the data, clean up the quantum number labels, and
return a performant `StructArray` containing the native `Transition`
types.

Idea is to also implement an overarching `Simulation` wrapper, which
will include the energy level parsing, too.